requirements.txt
streamlit>=1.34.0
pandas
openai
python-docx
pdfminer.six
psycopg2-binary
sqlalchemy
streamlit-cytoscapejs
